+++ index.ts
@@ -2,6 +2,13 @@ const hsvrgb = require("hsv2rgb");
 const rgbhsv = require("rgb-hsv");
 
 // REVIEW: https://www.typescriptlang.org/docs/handbook/2/functions.html#function-overloads
+
+/**
+ * convert normalized grayscale value to RGB
+ * @param val grayscale value 0..1
+ * @returns RGB color array[3] in 0..255
+ */
+export function hsv2rgb(val: number): number[];
 /**
  * convert normalized HSB color to RGB
  * @param arr [h, s, v] or [h, s, v, a]
@@ -27,28 +34,33 @@ export function hsv2rgb(h: number, s: number, v: number): number[];
 export function hsv2rgb(h: number, s: number, v: number, a: number): number[];
 /**
  * convert normalized HSB color to RGB
- * @param hOrArr normalized hue 0..1 (wrapped)
+ * @param hOrValOrArr normalized hue 0..1 (wrapped)
  * @param s normalized saturation 0..1 (clamped)
  * @param v normalized brightness 0..1 (clamped)
  * @param a alpha 0..1 passed-through
  * @returns RGB color array in 0..255
  */
-export function hsv2rgb(hOrArr: any, s?: number, v?: number, a?: number) {
+export function hsv2rgb(hOrValOrArr: any, s?: number, v?: number, a?: number) {
   if (s !== undefined && v !== undefined && a !== undefined) {
     // (h, s, v, a)
-    return [...hsvrgb(hOrArr * 360, s, v), a];
+    return [...hsvrgb(hOrValOrArr * 360, s, v), a];
   } else if (s !== undefined && v !== undefined) {
     // (h, s, v)
-    return [...hsvrgb(hOrArr * 360, s, v)];
+    return [...hsvrgb(hOrValOrArr * 360, s, v)];
   } else if (s === undefined) {
-    hOrArr[0] *= 360;
-    if (hOrArr.length === 3) {
-      // ([h, s, v])
-      return [...hsvrgb(...hOrArr)];
-    } else if (hOrArr.length === 4) {
-      // ([h, s, v, a])
-      // hsvrgb's 4th argument is NOT alpha but out array
-      return [...hsvrgb(...hOrArr.slice(0, 3)), hOrArr[3]];
+    if (Array.isArray(hOrValOrArr)) {
+      hOrValOrArr[0] *= 360;
+      if (hOrValOrArr.length === 3) {
+        // ([h, s, v])
+        return [...hsvrgb(...hOrValOrArr)];
+      } else if (hOrValOrArr.length === 4) {
+        // ([h, s, v, a])
+        // hsvrgb's 4th argument is NOT alpha but out array
+        return [...hsvrgb(...hOrValOrArr.slice(0, 3)), hOrValOrArr[3]];
+      }
+    } else {
+      // a single argument
+      return [...hsvrgb(0, 0, hOrValOrArr)];
     }
   }
   throw new Error(